Refactoring to ApiBase class and allow POST calls
